Transform list into vector for tabulated-list-entries
tabulated-list-entries is expected to be a list of lists in the
following format:

 ((id1 [field11 field12 field13 ...])
  (id2 [field21 field22 field23 ...])
  ...)
where every field is either a string or of the form

  (label . property)

Unfortunately, the single `vector' call insteadl of `apply' yields a
vector with a list and thus

 ((id1 [(field11 field12 field13 ...)])
  ...)

Therefore, the vector is interpreted as a single field with
   (field1 . field2 field3 ...),
as contents.

If by any chance our number of fields is even, the property "list"
will contain an odd number of elements and thus is improper, which
yields a very confusing error down from the bowels of Emacs, as
property list handling is defined in C, not in Emacs Lisp.

This commit uses `apply` to transform the list of fields into a vector
of fields, and thus brings into the expected form for
tabulated-list-entries.